refactor: match interfaces to SQL templates
