--- src/lib/interfaces/types.ts
@@ -1,14 +1,0 @@
-export namespace Types {
-
-  export interface Type {
-    type_id: string
-    name: string
-    schema: string
-    format: string
-    description: string
-    enums: string[]
-  }
-  
-
-}
-
